Various fixes and improvements
