+++ src/Elders.Cronus.Projections.Cassandra/CassandraProvider.cs
@@ -1,0 +1,96 @@
+﻿using System;
+using Cassandra;
+using Elders.Cronus.Projections.Cassandra.ReplicationStrategies;
+using Microsoft.Extensions.Configuration;
+
+namespace Elders.Cronus.Projections.Cassandra
+{
+    public class CassandraProvider : ICassandraProvider
+    {
+        public const string ConnectionStringSettingKey = "cronus_projections_cassandra_connectionstring";
+
+        private readonly IConfiguration configuration;
+        protected readonly IKeyspaceNamingStrategy keyspaceNamingStrategy;
+        protected readonly ICassandraReplicationStrategy replicationStrategy;
+        protected readonly IInitializer initializer;
+
+        protected Cluster cluster;
+        protected ISession session;
+
+        private string baseConfigurationKeyspace;
+
+        public CassandraProvider(IConfiguration configuration, IKeyspaceNamingStrategy keyspaceNamingStrategy, ICassandraReplicationStrategy replicationStrategy, IInitializer initializer = null)
+        {
+            if (configuration is null) throw new ArgumentNullException(nameof(configuration));
+            if (keyspaceNamingStrategy is null) throw new ArgumentNullException(nameof(keyspaceNamingStrategy));
+            if (replicationStrategy is null) throw new ArgumentNullException(nameof(replicationStrategy));
+
+            this.configuration = configuration;
+            this.keyspaceNamingStrategy = keyspaceNamingStrategy;
+            this.replicationStrategy = replicationStrategy;
+            this.initializer = initializer;
+        }
+
+        public Cluster GetCluster()
+        {
+            if (cluster is null == false)
+                return cluster;
+
+            Builder builder = initializer as Builder;
+            if (builder is null)
+            {
+                builder = Cluster.Builder();
+                //  TODO: check inside the `cfg` (var cfg = builder.GetConfiguration();) if we already have connectionString specified
+
+                string connectionString = configuration.GetRequired(GetConnectionStringSettingKey());
+
+                var hackyBuilder = new CassandraConnectionStringBuilder(connectionString);
+                if (string.IsNullOrEmpty(hackyBuilder.DefaultKeyspace) == false)
+                    connectionString = connectionString.Replace(hackyBuilder.DefaultKeyspace, "");
+                baseConfigurationKeyspace = hackyBuilder.DefaultKeyspace;
+
+                var connStrBuilder = new CassandraConnectionStringBuilder(connectionString);
+                cluster = connStrBuilder
+                    .ApplyToBuilder(builder)
+                    .WithReconnectionPolicy(new ExponentialReconnectionPolicy(100, 100000))
+                    .WithRetryPolicy(new NoHintedHandOffRetryPolicy())
+                    .Build();
+            }
+
+            else
+            {
+                cluster = Cluster.BuildFrom(initializer);
+            }
+
+            return cluster;
+        }
+
+        protected virtual string GetConnectionStringSettingKey()
+        {
+            return ConnectionStringSettingKey;
+        }
+
+        protected virtual string GetKeyspace()
+        {
+            return keyspaceNamingStrategy.GetName(baseConfigurationKeyspace).ToLower();
+        }
+
+        public ISession GetSession()
+        {
+            if (session is null)
+            {
+                session = GetCluster().Connect();
+                CreateKeyspace(replicationStrategy, GetKeyspace());
+            }
+
+            return session;
+        }
+
+        private void CreateKeyspace(ICassandraReplicationStrategy replicationStrategy, string keyspace)
+        {
+            var createKeySpaceQuery = replicationStrategy.CreateKeySpaceTemplate(keyspace);
+            session.Execute(createKeySpaceQuery);
+            session.ChangeKeyspace(keyspace);
+        }
+    }
+}
