+++ src/Elders.Cronus.Projections.Cassandra/ICassandraProjectionStoreSettings.cs
@@ -1,0 +1,9 @@
+﻿namespace Elders.Cronus.Projections.Cassandra
+{
+    public interface ICassandraProjectionStoreSettings
+    {
+        ICassandraProvider CassandraProvider { get; }
+        ISerializer Serializer { get; }
+        IProjectionsNamingStrategy ProjectionsNamingStrategy { get; }
+    }
+}
