+++ src/Elders.Cronus.Projections.Cassandra/IKeyspaceNamingStrategy.cs
@@ -1,0 +1,7 @@
+﻿namespace Elders.Cronus.Projections.Cassandra
+{
+    public interface IKeyspaceNamingStrategy
+    {
+        string GetName(string baseConfigurationKeyspace);
+    }
+}
