+++ src/Elders.Cronus.Projections.Cassandra/ITableNamingStrategy.cs
@@ -1,0 +1,7 @@
+﻿namespace Elders.Cronus.Projections.Cassandra
+{
+    public interface ITableNamingStrategy
+    {
+        string GetName();
+    }
+}
