+++ src/Elders.Cronus.Projections.Cassandra/NoKeyspaceNamingStrategy.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace Elders.Cronus.Projections.Cassandra
+{
+    public sealed class NoKeyspaceNamingStrategy : IKeyspaceNamingStrategy
+    {
+        public string GetName(string baseConfigurationKeyspace)
+        {
+            if (baseConfigurationKeyspace.Length > 48) throw new ArgumentException($"Cassandra keyspace exceeds maximum length of 48. Keyspace: {baseConfigurationKeyspace}");
+
+            return baseConfigurationKeyspace;
+        }
+    }
+}
